Added a bunch of missing FL_EXTERN's to glut.h (I have not really tested this, but I believe this is why I keep getting mail about glut not linking on win32.  If anybody can check this please do so!)
Fix for sliders so that clicking on one with a small (or zero)
slider_size will not move the slider.

I put Fl_Button.H in fl_shortcut.cxx so that it gets the FL_EXTERN for
fl_old_shortcut() (and also assures the declarations match).

Fixed xpaint link in the documentation.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.0@1212 ea41ed52-d2ee-0310-a9c1-e6b18d33e121