Added Fl::option() setting to allow arrow keys to navigate table.
This example was written before 1.3, and was dependent on the 'old' FLTK behavior
where arrow keys could be used to navigate widget focus (in addition to Tab).

The 'new' FLTK behavior (1.3 and up) disables arrow key focus nav by default,
which is bad for this program, so we enable the old FLTK behavior with Fl::option().



git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@9838 ea41ed52-d2ee-0310-a9c1-e6b18d33e121