STR 3480: add "Missing method(s) in Fl_Check_Browser".
This is a complete proof of concept how to add the missing methods.

FL/Fl_Check_Browser.H: new methods (the only required change).
test/check_browser.cxx: test program to exercise the new methods.

Other files: build and git configuration.

Note: this is NOT intended to be committed to FLTK 1.3, it is ONLY
meant as a proof of concept. The new virtual methods break the ABI.

The final solution will be available in FLTK 1.4 and will presumably
not be header-only changes - the implementation of some of the new
methods will be in a library source file.